// src/app/api/models/tier-details.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Entitlements } from '../models/entitlements';
import { Pricing } from '../models/pricing';
import { Tier } from '../models/tier';
export interface TierDetails {
  entitlements: Entitlements;
  name: string;
  pricing: Pricing;
  tier: Tier;
}